refactor: Modularize application into separate files
