+++ config.py
@@ -1,0 +1,8 @@
+import os
+from dotenv import load_dotenv
+
+load_dotenv()
+
+API_KEY = os.getenv('GEMINI_API_KEY')
+API_URL = f"https://generativelanguage.googleapis.com/v1beta/models/gemini-1.5-pro-latest:generateContent?key={API_KEY}"
+REDIS_URL = os.getenv('REDIS_URL', 'redis://localhost:6379')
